refactor pharmacy order
