+++ packages/database/src/models/PharmacyOrderPrescription/index.ts
@@ -1,0 +1,1 @@
+export * from './PharmacyOrderPrescription';
